Refactor items (tower, upgradeitem) to be generic
Purchaseable
Is a generic abstract class for any purchaseable item.
Contains information about the item, for use in the shop.
Use .create() to return an Item

Item
Represents an instance of a shop Item, such as a Tower

UpgradeItem is a bit strange in that there aren't multiple instances of an upgradeitem.
So UpgradItem is both Purchasable and an Item.